PizzaWebApp/requirements.txt
Flask
markupsafe
flask_bcrypt
pyodbc
azure-keyvault-secrets
azure-identity
waitress